+++ Source/Rules/TrailblazerRule_CostAllowedArea.cs
@@ -6,13 +6,13 @@ namespace Trailblazer.Rules
     /// <summary>
     /// Cost penalty to a cell if it's outside the allowed area of the pawn
     /// </summary>
-    public abstract class TrailblazerRule_CostAllowedArea : TrailblazerRule
+    public class TrailblazerRule_CostAllowedArea : TrailblazerRule
     {
         protected const int Cost_OutsideAllowedArea = 600;
 
         protected readonly Area allowedArea;
 
-        TrailblazerRule_CostAllowedArea(PathData pathData) : base(pathData)
+        public TrailblazerRule_CostAllowedArea(PathData pathData) : base(pathData)
         {
             Pawn pawn = pathData.traverseParms.pawn;