+++ changelog.md
@@ -1,2 +1,4 @@
 # v1.0.0 - Initial Release
 **Features**
+ - Extensible pathfinding rule engine
+   - Other mods can easily add or modify pathfinding rules